requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py.git@934ab4151add79a2140c7084679b15b6c5eb6f91